Converting to config file
revising scripts to use config file and lower case variables because that is the right thing to do. Upper case vars have a special sort of traditional meaning and are used by the system. We dont like to break the system.
Rename fureverso to quack
rename transfer and unpack plex to have plex prefix